UI modularized custom pipes, enhanced threshold fields on alert forms
